+++ tests/skoolkittest.py
@@ -14,7 +14,8 @@ from unittest import TestCase
 
 SKOOLKIT_HOME = abspath(dirname(dirname(__file__)))
 sys.path.insert(0, SKOOLKIT_HOME)
-from skoolkit import bin2tap, skool2asm, skool2bin, skool2ctl, skool2html, skool2sft, sna2skool, tap2sna, tapinfo
+from skoolkit import (bin2sna, bin2tap, skool2asm, skool2bin, skool2ctl,
+                      skool2html, skool2sft, sna2skool, tap2sna, tapinfo)
 
 def get_parity(data):
     parity = 0
@@ -296,6 +297,9 @@ class SkoolKitTestCase(TestCase):
         out = self.to_lines(self.out.getvalue(), strip_cr) if out_lines else self.out.getvalue()
         err = self.to_lines(self.err.getvalue(), strip_cr) if err_lines else self.err.getvalue()
         return out, err
+
+    def run_bin2sna(self, args='', out_lines=True, err_lines=False, strip_cr=True, catch_exit=None):
+        return self._run_skoolkit_command(bin2sna.main, args, out_lines, err_lines, strip_cr, catch_exit)
 
     def run_bin2tap(self, args='', out_lines=True, err_lines=False, strip_cr=True, catch_exit=None):
         return self._run_skoolkit_command(bin2tap.main, args, out_lines, err_lines, strip_cr, catch_exit)