+++ sphinx/source/changelog.rst
@@ -3,6 +3,8 @@ Changelog
 
 5.2b1
 -----
+* Added the :ref:`bin2sna.py` command (for converting a binary file into a Z80
+  snapshot)
 * Added the :ref:`N` macro (which renders a numeric value in hexadecimal format
   when the ``--hex`` option is used with `skool2asm.py` or `skool2html.py`)
 * When the :ref:`LINK` macro links to an entry on a memory map page, the anchor